Fixed file references
changed instances of com.example.sanat to au.com.knight_fighters
