+++ KNIGHTFIGHTERS/app/src/main/java/au/com/knight_fighters/Level_3/Level3.java
@@ -1,22 +1,23 @@
-package com.example.sanat.knightfighters;
+package au.com.knight_fighters.Level_3;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
+import au.com.knight_fighters.R;
 
-public class Level2 extends AppCompatActivity {
+public class Level3 extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_level2);
+        setContentView(R.layout.activity_level3);
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_level2, menu);
+        getMenuInflater().inflate(R.menu.menu_level3, menu);
         return true;
     }
