--- KNIGHTFIGHTERS/app/src/main/java/com/example/sanat/knightfighters/Level4background.java
@@ -1,63 +1,0 @@
-package com.example.sanat.knightfighters;
-
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuItem;
-import android.view.MotionEvent;
-import android.view.View;
-import android.widget.RelativeLayout;
-
-public class Level4background extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_level4background);
-        RelativeLayout levelbg = (RelativeLayout)findViewById(R.id.levelbg);
-
-        levelbg.setOnTouchListener(
-                new RelativeLayout.OnTouchListener() {
-                    public boolean onTouch(View v, MotionEvent m) {
-
-                        handletouch(m);
-
-                        return true;
-
-                    }
-                }
-        );
-    }
-    public void handletouch(MotionEvent m){
-        if(m.getActionMasked() == MotionEvent.ACTION_DOWN){
-            startActivity(new Intent(this, Level4.class));
-            finish();
-
-        }
-
-
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_level4background, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-}
