+++ KNIGHTFIGHTERS/app/src/main/res/layout/activity_level3.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.example.sanat.knightfighters.Level3">
+    tools:context="au.com.knight_fighters.Level_3.Level3">
 
     <TextView
         android:layout_width="wrap_content"
