+++ KNIGHTFIGHTERS/app/src/main/res/menu/menu_level2background.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.sanat.knightfighters.Level2background">
+    tools:context="au.com.knight_fighters.Level_2.Level2background">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" app:showAsAction="never" />
 </menu>
