+++ KNIGHTFIGHTERS/app/src/main/res/menu/menu_level3.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.example.sanat.knightfighters.Level3">
+    tools:context="au.com.knight_fighters.Level_3.Level3">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" app:showAsAction="never" />
 </menu>
